Refactoring. Now have build error with network DI
