+++ .idea/gradle.xml
@@ -11,10 +11,12 @@
             <option value="$PROJECT_DIR$" />
             <option value="$PROJECT_DIR$/app" />
             <option value="$PROJECT_DIR$/feature" />
+            <option value="$PROJECT_DIR$/feature/app-ui" />
             <option value="$PROJECT_DIR$/feature/arch" />
             <option value="$PROJECT_DIR$/feature/localstorage" />
             <option value="$PROJECT_DIR$/shared" />
             <option value="$PROJECT_DIR$/shared/api" />
+            <option value="$PROJECT_DIR$/shared/ui-kit" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
