--- app/src/main/java/com/sedooj/resumen/navigation/graph.kt
@@ -1,11 +1,0 @@
-package com.sedooj.resumen.navigation
-
-//@NavHostGraph(
-//    route = "auth_route"
-//)
-//annotation class AuthorizationGraph
-//
-//@NavHostGraph(
-//    route = "home_route"
-//)
-//annotation class HomeGraph
