--- app/src/main/java/com/sedooj/resumen/network/connectionHandler/ConnectionHandler.kt
@@ -1,80 +1,0 @@
-package com.sedooj.resumen.network.connectionHandler
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.Network
-import android.net.NetworkCapabilities
-import android.net.NetworkRequest
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.State
-import androidx.compose.runtime.produceState
-import androidx.compose.ui.platform.LocalContext
-import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.flow.callbackFlow
-
-
-sealed class ConnectionState {
-    object Available : ConnectionState()
-    object Unavailable : ConnectionState()
-}
-class ConnectionHandler {
-}
-
-val Context.currentConnectivityState: ConnectionState
-    get() {
-        val connectivityManager =
-            getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        return getCurrentConnectivityState(connectivityManager)
-    }
-
-private fun getCurrentConnectivityState(
-    connectivityManager: ConnectivityManager
-): ConnectionState {
-    val connected = connectivityManager.allNetworks.any { network ->
-        connectivityManager.getNetworkCapabilities(network)
-            ?.hasCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-            ?: false
-    }
-
-    return if (connected) ConnectionState.Available else ConnectionState.Unavailable
-}
-
-@ExperimentalCoroutinesApi
-fun Context.observeConnectivityAsFlow() = callbackFlow {
-    val connectivityManager = getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-
-    val callback = NetworkCallback { connectionState -> trySend(connectionState) }
-
-    val networkRequest = NetworkRequest.Builder()
-        .addCapability(NetworkCapabilities.NET_CAPABILITY_INTERNET)
-        .build()
-
-    connectivityManager.registerNetworkCallback(networkRequest, callback)
-    val currentState = getCurrentConnectivityState(connectivityManager)
-    trySend(currentState)
-    awaitClose {
-        connectivityManager.unregisterNetworkCallback(callback)
-    }
-}
-
-fun NetworkCallback(callback: (ConnectionState) -> Unit): ConnectivityManager.NetworkCallback {
-    return object : ConnectivityManager.NetworkCallback() {
-        override fun onAvailable(network: Network) {
-            callback(ConnectionState.Available)
-        }
-
-        override fun onLost(network: Network) {
-            callback(ConnectionState.Unavailable)
-        }
-    }
-}
-
-@ExperimentalCoroutinesApi
-@Composable
-fun connectivityState(): State<ConnectionState> {
-    val context = LocalContext.current
-    return produceState(initialValue = context.currentConnectivityState) {
-        context.observeConnectivityAsFlow().collect { value = it }
-    }
-}
