--- app/src/main/java/com/sedooj/resumen/pages/Routes.kt
@@ -1,9 +1,0 @@
-package com.sedooj.resumen.pages
-object Routes {
-    const val SIGN_IN = "SIGN_IN"
-    const val SIGN_IN_CONFIRMATION = "SIGN_IN_CONFIRMATION"
-    const val SIGN_UP = "SIGN_UP"
-    const val HOME = "HOME"
-    const val MY_RESUMES = "MY_RESUMES"
-    const val PROFILE = "PROFILE"
-}
