+++ feature/app-ui/build.gradle.kts
@@ -1,34 +1,27 @@
-@Suppress("DSL_SCOPE_VIOLATION") // TODO: Remove once KTIJ-19369 is fixed
 plugins {
-    alias(libs.plugins.androidApplication)
+    alias(libs.plugins.androidLibrary)
     alias(libs.plugins.kotlinAndroid)
+    alias(libs.plugins.ksp)
 }
 
 android {
-    namespace = "com.sedooj.resumen"
+    namespace = "com.sedooj.app_ui"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.sedooj.resumen"
         minSdk = 29
-        targetSdk = 34
-        versionCode = 1
-        versionName = "1.0"
 
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-        vectorDrawables {
-            useSupportLibrary = true
-        }
+        consumerProguardFiles("consumer-rules.pro")
     }
 
     buildTypes {
         release {
-            isMinifyEnabled = true
+            isMinifyEnabled = false
             proguardFiles(
                 getDefaultProguardFile("proguard-android-optimize.txt"),
                 "proguard-rules.pro"
             )
-            signingConfig = signingConfigs.getByName("debug")
         }
     }
     compileOptions {
@@ -44,19 +37,17 @@ android {
     composeOptions {
         kotlinCompilerExtensionVersion = "1.5.10"
     }
-    packaging {
-        resources {
-            excludes += "/META-INF/{AL2.0,LGPL2.1}"
-        }
-    }
 }
 
 dependencies {
-    implementation(libs.slf4j)
-    implementation(project(":feature:arch"))
-    implementation(libs.androidx.core.splashscreen)
-    api(libs.androidx.appcompat)
     api(libs.core.ktx)
+    implementation(project(":feature:arch"))
+    ksp(libs.destinations.ksp)
+    implementation(libs.destinations.core)
+    implementation(project(":shared:ui-kit"))
+    implementation(project(":shared:api"))
+
+
     api(libs.material)
     api(libs.activity.compose)
     api(platform(libs.compose.bom))
@@ -64,14 +55,8 @@ dependencies {
     api(libs.ui.graphics)
     api(libs.ui.tooling.preview)
     api(libs.material3)
-    implementation(project(":feature:app-ui"))
-
 
     testImplementation(libs.junit)
     androidTestImplementation(libs.androidx.test.ext.junit)
     androidTestImplementation(libs.espresso.core)
-    androidTestImplementation(platform(libs.compose.bom))
-    androidTestImplementation(libs.ui.test.junit4)
-    debugImplementation(libs.ui.tooling)
-    debugImplementation(libs.ui.test.manifest)
 }
