+++ feature/app-ui/src/main/java/com/sedooj/app_ui/navigation/graph.kt
@@ -1,0 +1,11 @@
+package com.sedooj.app_ui.navigation
+
+//@NavHostGraph(
+//    route = "auth_route"
+//)
+//annotation class AuthorizationGraph
+//
+//@NavHostGraph(
+//    route = "home_route"
+//)
+//annotation class HomeGraph
