+++ settings.gradle.kts
@@ -18,3 +18,5 @@ include(":app")
 include(":feature:arch")
 include(":feature:localstorage")
 include(":shared:api")
+include(":shared:ui-kit")
+include(":feature:app-ui")
