+++ shared/ui-kit/src/main/res/drawable/home.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+  <path
+      android:pathData="m498.2,222.7 l-0.04,-0.04L289.3,13.81C280.4,4.91 268.57,0 255.98,0c-12.59,0 -24.43,4.9 -33.33,13.81L13.9,222.55c-0.07,0.07 -0.14,0.14 -0.21,0.22 -18.28,18.39 -18.25,48.22 0.09,66.56 8.38,8.38 19.44,13.24 31.28,13.75 0.48,0.05 0.96,0.07 1.45,0.07h8.32v153.7C54.83,487.25 79.58,512 110,512h81.71c8.28,0 15,-6.72 15,-15L206.71,376.5c0,-13.88 11.29,-25.17 25.17,-25.17h48.19c13.88,0 25.17,11.29 25.17,25.17L305.24,497c0,8.28 6.72,15 15,15h81.71c30.42,0 55.17,-24.75 55.17,-55.16v-153.7h7.72c12.59,0 24.42,-4.9 33.33,-13.81 18.36,-18.37 18.37,-48.25 0.02,-66.64zM498.2,222.7"
+      android:fillColor="#000000"/>
+</vector>
