+++ shared/ui-kit/src/main/res/drawable/my_resumes.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="408.02"
+    android:viewportHeight="408.02">
+  <path
+      android:pathData="M367.73,112.65H40.29c-22.27,0.13 -40.26,18.21 -40.28,40.48a5.14,5.14 0,0 0,0 0.68l23.4,174.6c0.28,22.16 18.32,39.98 40.48,40h280.4c22.16,-0.02 40.2,-17.84 40.48,-40l23.24,-174.6a5.14,5.14 0,0 0,0 -0.68c-0.02,-22.27 -18.01,-40.35 -40.28,-40.48zM337.85,72.33h-131.52l-26,-30.92a5,5 0,0 0,-4 -1.8H70.17c-16.56,0.02 -29.98,13.44 -30,30v28.84h327.4c-1.95,-14.93 -14.66,-26.11 -29.72,-26.12z"
+      android:fillColor="#000000"/>
+</vector>
