+++ shared/ui-kit/src/main/res/drawable/no_internet_connection.xml
@@ -1,0 +1,12 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="512dp"
+    android:height="512dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M11.997,18.5m-1.5,0a1.5,1.5 0,1 1,3 0a1.5,1.5 0,1 1,-3 0"
+      android:fillColor="#000000"/>
+  <path
+      android:pathData="M2.307,9.548a1,1 0,0 0,1.414 -0.022C5.934,7.242 8.874,5.986 12,5.984c0.98,0 1.938,0.133 2.863,0.375L13.44,8.492A8.906,8.906 0,0 0,12 8.36c-2.54,0 -4.924,1.019 -6.716,2.867A1,1 0,0 0,6.72 12.62c1.412,-1.457 3.287,-2.26 5.28,-2.26 0.062,0 0.123,0.013 0.185,0.015l-1.762,2.643c-0.798,0.25 -1.534,0.679 -2.14,1.304a1,1 0,0 0,0.022 1.414,0.986 0.986,0 0,0 0.207,0.148l-3.133,4.7a0.75,0.75 0,1 0,1.248 0.832l12,-18a0.75,0.75 0,1 0,-1.248 -0.832l-1.36,2.04a13.256,13.256 0,0 0,-4.02 -0.64c-3.67,0.001 -7.121,1.475 -9.714,4.15a1,1 0,0 0,0.022 1.414zM20.997,9.83a1,1 0,0 0,0.718 -1.696,13.735 13.735,0 0,0 -3.436,-2.572l-1.117,1.676a11.705,11.705 0,0 1,3.117 2.288,0.998 0.998,0 0,0 0.718,0.304zM17.998,12.924a1,1 0,0 0,0.718 -1.696,9.461 9.461,0 0,0 -2.866,-2.023l-1.13,1.697c0.945,0.382 1.819,0.954 2.56,1.718a0.998,0.998 0,0 0,0.718 0.304zM14.999,16.018a1,1 0,0 0,0.718 -1.696,5.147 5.147,0 0,0 -2.358,-1.38l-1.228,1.843c0.79,0.03 1.571,0.332 2.15,0.93a0.998,0.998 0,0 0,0.718 0.303z"
+      android:fillColor="#000000"/>
+</vector>
