+++ shared/ui-kit/src/main/res/drawable/splash_screen_hand.xml
@@ -1,0 +1,68 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="100dp"
+    android:height="100dp"
+    android:autoMirrored="true"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+
+    <path
+        android:fillColor="#F9BC87"
+        android:pathData="M132.9,149.2l-28.6,-44.1c-3.9,-6 -2.1,-14.1 4,-17.9L170,48.5c6,-3.8 13.9,-2 17.7,4l32.7,51.4l57.3,73.3l-131.8,62.3l-11.9,-41.7c-2.1,-7.2 -2.6,-14.8 -1.4,-22.3l2.5,-16.4C135.6,155.7 134.8,152.1 132.9,149.2L132.9,149.2z" />
+
+    <path
+        android:fillColor="#F6A96D"
+        android:pathData="M284.8,181.5l-14.7,-39.6c-2.2,-5.9 -6.3,-11 -11.7,-14.3l-38.1,-23.6l-32.7,-51.4c-3.8,-6 -11.7,-7.8 -17.7,-4l-3.1,1.9c0.5,0.5 23,35.8 31.3,48.8c2,3.1 4.7,5.7 7.8,7.7l33.3,20.7c5.4,3.3 9.5,8.4 11.7,14.3l14.7,39.6c2,5.5 0.4,11.4 -3.6,15.1c3.8,3.5 9.3,4.8 14.3,3C283.8,196.9 287.5,188.8 284.8,181.5L284.8,181.5z" />
+
+    <path
+        android:fillColor="#F6A96D"
+        android:pathData="M231.7,170.6l-4.3,-5.3l-24.7,-0.8c-11.6,-0.4 -22.2,-5.9 -29.2,-15.1l-16.6,-21.9c-1.5,-2 -1.1,-4.8 0.9,-6.3c2,-1.5 4.8,-1.1 6.3,0.9l0,0l16.6,21.9c5.3,7.1 13.5,11.3 22.4,11.5l26.9,0.8c1.4,0 2.6,0.7 3.5,1.8l7.5,9.9L231.7,170.6z" />
+
+    <path
+        android:fillColor="#F6A96D"
+        android:pathData="M151.7,273.4" />
+
+    <path
+        android:fillColor="#E9834D"
+        android:pathData="M286.5,178.1c-1,-0.2 -10.2,-1.8 -24.9,-0.6c-18.9,1.5 -38.8,6.8 -57.7,15.5c-18.9,8.7 -35.9,20.3 -49.3,33.7C144.3,237 139.5,245 139,245.9c-2.6,4.4 -2.9,9.6 -0.9,14l10.2,22.1l18.8,0.5L179,268l-6.5,-14.2c6.6,-7.8 20.7,-21.7 44.5,-32.6c23.8,-11 43.5,-12.6 53.7,-12.5l6.5,14.2l18.8,0.5l11.9,-14.6l-10.2,-22.1C295.7,182.1 291.6,179 286.5,178.1L286.5,178.1z" />
+
+    <path
+        android:fillColor="#D76F41"
+        android:pathData="M146,276.9l6.8,14.9l18.7,0.1l12,-14.3l-6.8,-14.9L146,276.9zM305.7,203.5L275,217.6l6.8,14.9l18.7,0.1l12,-14.3L305.7,203.5z" />
+
+    <path
+        android:fillColor="#385F7E"
+        android:pathData="M149,384.5l31.3,68.1c3.7,8.1 13.9,11.4 22.7,7.3l208.8,-96c8.8,-4 12.9,-13.9 9.2,-22l-31.3,-68.1L149,384.5z" />
+
+    <path
+        android:fillColor="#2C4E67"
+        android:pathData="M145.1,376l3.9,8.5c17.7,3.6 61.2,6.2 133.4,-26.9s98.5,-68 107.2,-83.7l-3.9,-8.5L145.1,376L145.1,376z" />
+
+    <path
+        android:fillColor="#427194"
+        android:pathData="M142.1,371.5c17,3.7 60.9,7 134.3,-26.7s99.4,-69.2 107.7,-84.5c2.4,-4.5 0.8,-13.7 -1.2,-18.2l-18.8,-40.9c-3.7,-8.1 -13.9,-11.4 -22.7,-7.3l-221.9,102c-8.8,4 -12.9,13.9 -9.2,22l18.8,40.9C131.2,363.1 137.1,370.4 142.1,371.5z" />
+
+    <path
+        android:fillColor="#385F7E"
+        android:pathData="M382.2,256.1c-8.3,15.3 -34.3,50.8 -107.7,84.5S157.3,371 140.2,367.4c-5,-1.1 -9.1,-4.2 -11.1,-8.7l3.9,8.5c2,4.4 6.1,7.6 11.1,8.7c17,3.7 60.9,7 134.3,-26.7s99.4,-69.2 107.7,-84.5c2.4,-4.5 2.7,-9.6 0.7,-14.1l-3.9,-8.5C384.9,246.5 384.7,251.6 382.2,256.1z" />
+
+    <path
+        android:fillColor="#FFE27B"
+        android:pathData="M284.9,363.2c8.2,-3.8 12.3,-12.5 6.7,-24.7l-3.8,-8.3c-1.2,-2.7 -4.6,-3.8 -7.6,-2.4l-19.2,8.8c-2.9,1.3 -4.3,4.6 -3.1,7.3l3.8,8.3C267.4,364.4 276.7,367 284.9,363.2z" />
+
+    <path
+        android:fillColor="#F9D05A"
+        android:pathData="M283,359.1c-8.2,3.8 -17.7,0.7 -21.2,-6.8l3.9,8.5c3.5,7.6 13,10.6 21.2,6.8s12.1,-13 8.6,-20.6l-3.9,-8.5C295.1,346.1 291.2,355.3 283,359.1L283,359.1z" />
+
+    <path
+        android:fillColor="#F9BC87"
+        android:pathData="M265.5,182c-4.1,-8 -17.7,-6.8 -17.7,-6.8c-6.3,4.6 -8.6,13.3 -4.9,20.5l-28.1,-13.8c-6.3,4.6 -8.5,13.3 -4.9,20.5l0.4,0.8c-4.1,-8 -24,-5.7 -24,-5.7c-6.2,4.6 -8.5,13.3 -4.9,20.5l10.6,20.8c-9,-5.1 -23.9,-5.5 -23.9,-5.5c-6.3,4.6 -8.6,13.3 -4.9,20.5l7.1,14c4.1,9.8 9.2,4.1 17.2,0c8,-4.1 15.8,-6.8 11.7,-14.8c4.1,8 9.5,4.4 17.5,0.3c8,-4.1 15.5,-7.1 11.4,-15.1c4.1,8 7.8,7.7 15.8,3.7c8,-4.1 17.2,-10.4 13.2,-18.4c4.1,8 10.1,8 18.1,4C283.2,223.3 265.5,182 265.5,182L265.5,182z" />
+
+    <path
+        android:fillColor="#F6A96D"
+        android:pathData="M228.1,238.2c0,-0.1 0,-0.1 0,-0.2l-17.8,-34.9c-4.1,-8 -13.8,-11.2 -21.8,-7.1c-0.7,0.3 -1.3,0.7 -2,1.2c-0.1,0.1 -0.1,0.3 0,0.4c1.9,1.5 3.5,3.4 4.6,5.6l17.8,35c3.4,6.8 1.7,14.8 -3.8,19.6c-0.2,0.2 0,0.5 0.2,0.4c7.1,-2.7 6.7,-2.5 8.9,-3.6c5,-2.6 14.2,-5.6 15.1,-10.9L228.1,238.2zM257.1,223.4l-18.2,-35.8c-4.1,-8 -13.8,-11.2 -21.8,-7.1c-0.7,0.3 -1.3,0.7 -2,1.2c-0.1,0.1 -0.1,0.3 0,0.4c1.9,1.5 3.5,3.4 4.6,5.6l18.2,35.8c3.6,7.1 1.5,15.7 -4.7,20.4c-0.1,0.1 -0.1,0.2 -0.1,0.3c1.8,3 7.3,-0.4 12.1,-3c0,0 0,0 0.1,0C253.2,236.9 261.1,231.4 257.1,223.4L257.1,223.4zM286,208.6l-14.1,-27.7c-4.1,-8 -13.8,-11.2 -21.8,-7.1c-0.7,0.3 -1.3,0.7 -2,1.2c-0.1,0.1 -0.1,0.3 0,0.4c1.9,1.5 3.5,3.4 4.6,5.6l14.1,27.7c3.6,7.1 1.5,15.7 -4.7,20.3c-0.1,0.1 -0.1,0.2 0,0.3c2,2.4 9.3,0.2 14.9,-2.6C284.9,222.6 290.1,216.6 286,208.6zM199.2,252.8l-7.1,-14c-4.1,-8 -13.8,-11.2 -21.8,-7.1c-0.7,0.3 -1.3,0.7 -2,1.2c-0.1,0.1 -0.1,0.3 0,0.4c1.9,1.5 3.5,3.4 4.6,5.6l7.1,14c3.1,6.1 2,13.2 -2.2,18c-0.4,0.5 -0.1,2.1 0.4,1.8c19.3,-10.3 4.7,-2.7 10,-5.5C196.3,263.1 203.3,260.8 199.2,252.8L199.2,252.8z" />
+
+    <path
+        android:fillColor="#F9BC87"
+        android:pathData="M220.4,103.9l38.1,23.6c5.4,3.3 9.5,8.4 11.7,14.3l14.7,39.6c2.7,7.3 -1.1,15.4 -8.4,18.1c-5.8,2.1 -12.3,0.1 -16,-4.9l-25,-33.8l-26.9,-0.8c-10.2,-0.3 -19.7,-5.2 -25.8,-13.3l-16.6,-21.9L220.4,103.9z" />
+
+</vector>
